+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 2.4.3
+- Release Pumpkin Warden from Pumpkin Burrow when broken by TNT, Creeper, or Wither
 
 # 2.4.2
 - Fix Forgotten Village Terminators